Added comment explaining change to integration
Comment explains why simple summing is appropriate for integrating over the spectrum for the Chianti spectrum file currently included in xrtpy in which the emission is binned and mentions how the current version Chianti (at least Chiantipy) outputs a spectrum that includes line widths. Thus if the spectrum file is updated, it may be appropriate to revert back to Simpson's method.